adicoes de métodos de calculo e afins

+++ src/main/java/com/example/projetonutricaoback/models/Preparacao.java
@@ -30,7 +30,7 @@ public class Preparacao {
 
 	private double CustoTotal;
 
-	private int numPorcoes;
+	private double numPorcoes;
 
 	private double CustoPerCapta;
